Move rest of the techs to /v2 directory
