+++ lib/techs/v2/level-proto.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var U = require('../util'),
+var U = require('../../util'),
     PATH = require('path');
 
 exports.baseTechPath = require.resolve('./level.js');
