+++ lib/techs/v2/level.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var BEM = require('../..');
+var BEM = require('../../..');
 
 exports.API_VER = 2;
