+++ lib/techs/v2/project.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var U = require('../util'),
+var U = require('../../util'),
     Q = require('q'),
     PATH = require('path');
