+++ lib/techs/v2/tech-docs.js
