+++ lib/util.js
@@ -532,19 +532,34 @@ exports.arrayReverse = function(arr) {
 exports.getBemTechPath = function(name, opts) {
     opts = _.defaults(opts || {}, {
         throwWhenUnresolved: false,
-        version: 1
     });
 
     var bemLib = process.env.COVER? 'bem/lib-cov/' : 'bem/lib/',
-        bemTechs = PATH.unixToOs(bemLib +
-                                    (opts.version === 1 ? 'techs'
-                                                        : 'techs/v2')),
-        path = PATH.join(bemTechs, name);
+        v1Path = PATH.join(bemLib, 'techs', name),
+        v2Path = PATH.join(bemLib, 'techs', 'v2', name),
+        paths;
 
-    if(exports.isRequireable(path + '.js')) return path + '.js';
-    if(exports.isRequireable(path)) return path;
+    if (typeof opts.version === 'undefined') {
+        paths = [
+            v1Path,
+            v2Path
+        ];
+    } else if (opts.version === 1) {
+        paths = [
+            v1Path
+        ];
+    } else {
+        paths = [
+            v2Path
+        ];
+    }
 
-    if (!opts.throwWhenUnresolved) return PATH.unixToOs(bemLib + 'tech');
+    for (var i=0, path; path = paths[i]; i++) {
+        if(exports.isRequireable(path + '.js')) return path + '.js';
+        if(exports.isRequireable(path)) return path;
+    }
+
+    if (!opts.throwWhenUnresolved) return PATH.join(bemLib, 'tech');
 
     throw new Error('Unable to resolve "' + name + '" tech');
 };
