+++ test/util.js
@@ -32,6 +32,11 @@ describe('util', function() {
                          PATH.unixToOs(bemLib + 'techs/v2/css.js'));
         });
 
+        it("'bemdecl.js' resolves to 'bem/lib/techs/v2/bemdecl.js.js'", function () {
+            assert.equal(U.getBemTechPath('bemdecl.js'),
+                        PATH.unixToOs(bemLib + 'techs/v2/bemdecl.js.js'));
+        });
+
         it("'custom' resolves to 'bem/lib/tech'", function() {
             assert.equal(U.getBemTechPath('custom'), PATH.unixToOs(bemLib + 'tech'));
         });
